% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{assoc}
\alias{assoc}
\alias{assoc,GPATree-method}
\title{Association mapping}
\usage{
\S4method{assoc}{GPATree}(object, FDR = 0.05, fdrControl = "global")
}
\arguments{
\item{object}{An object of class GPATree.}

\item{FDR}{FDR level. Value has to be between 0 and 1.}

\item{fdrControl}{Method to control FDR. Possible values are "global" (global FDR control) and "local" (local FDR control).}
}
\value{
Returns a binary matrix indicating the association between SNPs and phenotype, where its rows match those of input pvalue matrix for function GPATREE.
}
\description{
This function will implement association mapping for the GPA-Tree model.
}
\author{
Aastha Khatiwada
}
