% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decTree.R
\name{decTree}
\alias{decTree}
\alias{decTree,GPATree-method}
\title{GPATree selected decision tree}
\usage{
\S4method{decTree}{GPATree}(object)
}
\arguments{
\item{object}{An object of class GPATree.}
}
\value{
A list containing variables in combinations selected by the decision tree (CART_PIs), the combination (CART_PIs_comb) and the predicted proportions for the selected PIs(assoc_pred) meet the provided threshold in minPredictedProb.
}
\description{
This function will extract the combinations of functional annotations selected by the decision tree in the stage 2 of GPATree method that meets the provided threshold in minPredictedProb.
}
\author{
Aastha Khatiwada
}
