% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPATree.R
\name{GPATree}
\alias{GPATree}
\title{Implement the GPA-Tree approach}
\usage{
GPATree(gwasPval, annMat, initAlpha = 0.1, cpTry = 0.001)
}
\arguments{
\item{gwasPval}{A matrix of M X 1 dimension, where M is the number of SNPs. The matrix includes the GWAS association p-values for the phenotype. P-values must be between 0 and 1.}

\item{annMat}{A matrix of binary annotations, where row and column correspond to SNPs and annotations, respectively.}

\item{initAlpha}{Initial value for alpha estimate. Default is 0.1.}

\item{cpTry}{Complexity parameter (cp) value to be used. cpTry can be between 0 and 1 or NULL. Default is 0.001. When cpTry is NULL, GPATree will select the optimal cp to be used.}
}
\value{
This function returns a \code{List} including:
\itemize{
\item numIterConvergence: number of iterations taken for GPA-Tree approach to converge in Stage 1 and 2.
\item alpha: alpha estimated by the GPA-Tree approach.
\item fit: GPA-Tree model fit.
\item fitSelectVar: annotations included in the tree.
\item Z: posterior probability of being null and non-null for all SNPs.
\item zMarg: posterior probability of being a non-null SNP.
\item pi: updated prior based on GPA-Tree model.
\item licVec: vector of incomplete log-likelihood for all iteration of GPA-Tree approach.
\item lcVec: list of complete log likelihood for all iteration of GPA-Tree approach.
\item annMat: annMat provided by user.
\item gwasPval: gwasPval provided by user.
}
}
\description{
This function will implement the GPA-Tree approach.
}
\author{
Aastha Khatiwada
}
