% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPATreeExampleData.R
\docType{data}
\name{GPATreeExampleData}
\alias{GPATreeExampleData}
\alias{GPATreeExampleData,GPATree-method}
\title{GPATreeExampleData}
\format{
A List that contains 2 data frames.
\itemize{ The first data frame contains 100000 rows and 1 column. The rows represent SNPs and the columns represent the GWAS association p-value for the association between the SNP and some phenotype
}
\itemize{ The second dataframe contains 100000 rows and 10 columns. The rows represent SNPs and the columns represent 10 separate binary functional annotations.
}
}
\source{
Simulated Data
}
\usage{
GPATreeExampleData
}
\description{
Simulated data for the GPATree package
}
\keyword{datasets}
