% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPATreeStage1.R
\name{GPATreeStage1}
\alias{GPATreeStage1}
\title{Implement the Stage 1 of the GPA-Tree Method}
\usage{
GPATreeStage1(gwasPval, annMat, initAlpha = 0.1)
}
\arguments{
\item{gwasPval}{A matrix of M X 1 dimension where M is the number of SNPs. The matrix contains GWAS association p-values. Values must be between 0 and 1.}

\item{annMat}{A matrix of binary annotations, where row and column correspond to SNPs and annotations, respectively.}

\item{initAlpha}{Initial value for alpha estimate. Default is 0.1.}
}
\value{
This function returns a \code{List} including:
\itemize{
\item num_iter_stage1: number of iterations taken for Stage 1 of GPA-Tree to converge.
\item pi_stage1: predicted posterior probability of being a non-null SNP in Stage 1 of GPA-Tree Method.
\item alpha_stage1: estimated alpha of GPA-Tree Method.
\item beta_stage1: beta parameters from the linear model fitted at convergence of Stage 1 of GPA-Tree Method.
}
}
\description{
This function will implement the stage 1 of the GPA-Tree method.
}
\author{
Aastha Khatiwada
}
