% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPATreeStage2.R
\name{GPATreeStage2}
\alias{GPATreeStage2}
\title{Implement the Stage 2 of the GPA-Tree approach.}
\usage{
GPATreeStage2(gwasPval, annMat, alphaStage1, initPi, cpTry)
}
\arguments{
\item{gwasPval}{A matrix of M X 1 dimension, where M is the number of SNPs. The first column is labeled 'SNPid' and contains the SNPid. The second column contains the GWAS association p-values and is called 'P1'. The values in P1 must be between 0 and 1.}

\item{annMat}{A matrix of binary annotations, where row and column correspond to SNPs and annotations, respectively.}

\item{alphaStage1}{alpha estimated in stage 1 of the GPA-Tree approach.}

\item{initPi}{pi estimated in stage 1 of the GPA-Tree approach.}

\item{cpTry}{Complexity parameter (cp) value to be used to build annotation decision tree. cpTry can be between 0 and 1 or NULL. Default is 0.001. When cpTry is NULL, GPATree will select the optimal cp to be used.}
}
\value{
This function returns a \code{List} including:
\itemize{
\item num_iter_stage2: number of iterations taken for Stage 2 of GPA-Tree approach to converge.
\item lic_list_stage2: list of incomplete log-likelihood for all iterations in Stage 2 of GPA-Tree approach.
\item lc_list_stage2: list of complete log-likelihood for all iterations in Stage 2 of GPA-Tree approach.
\item zi_stage2: probability of being a null  and non-null SNP.
\item Zmarg: marginal posterior probability of being a non-null SNP for the phenotype.
\item pi_stage2: updated prior.
\item var_in_tree: annotations included in the tree.
\item pruned_tree: GPA-Tree model fit.
}
}
\description{
This function will implement the Stage 2 of the GPA-Tree approach.
}
\author{
Aastha Khatiwada
}
