% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf.R
\name{leaf}
\alias{leaf}
\alias{leaf,GPATree-method}
\title{Functional annotation tree.}
\usage{
\S4method{leaf}{GPATree}(object)
}
\arguments{
\item{object}{An object of class GPATree.}
}
\value{
Returns a matrix where each row corresponds to a leaf from the GPA-Tree model fit and contains information regarding the local FDR for SNPs in the leaf, and also information regarding annotations that are enriched (1) or not enriched (0) for the leaf.
}
\description{
This function will provide the annotation combinations relevant to risk-associated SNPs.
}
\examples{
\dontrun{
library(GPATree)

# load GPATree example data
data(GPATreeExampleData)

#fitting the GPATree model
fit <- GPATree(GPATreeExampleData$gwasPval, GPATreeExampleData$annMat)
leaf(fit)
}
}
\author{
Aastha Khatiwada
}
