% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot,GPATree,missing-method}
\title{Plot the functional annotation tree}
\usage{
\S4method{plot}{GPATree,missing}(x, y, ...)
}
\arguments{
\item{x}{An object of class GPATree.}

\item{y}{missing (not required).}

\item{...}{...}
}
\value{
Returns a plot for the functional annotation tree from the GPA-Tree model fit.
}
\description{
This function will plot the functional annotation tree for the GPA-Tree model fit.
}
\examples{
\dontrun{
library(GPATree)

# load GPATree example data
data(GPATreeExampleData)

#fitting the GPATree model
fit <- GPATree(GPATreeExampleData$gwasPval, GPATreeExampleData$annMat)

# plotting the GPATree model fit
plot(fit)
}
}
\author{
Aastha Khatiwada
}
