% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune}
\alias{prune}
\alias{prune,GPATree-method}
\title{Prune GPA-Tree model fit}
\usage{
\S4method{prune}{GPATree}(object, cp = 0.001)
}
\arguments{
\item{object}{An object of class GPATree.}

\item{cp}{The cp parameter to be used for pruning. cp must be between 0 and 1.}
}
\value{
GPA-Tree model output.
}
\description{
This function will prune the GPA-Tree model fit using the given cp value.
}
\examples{
\dontrun{
library(GPATree)

# load GPATree example data
data(GPATreeExampleData)

#fitting the GPATree model
fit <- GPATree(GPATreeExampleData$gwasPval, GPATreeExampleData$annMat)

# pruning the GPATree model fit
pruned.fit <- prune(fit, cp = 0.005)
}
}
\author{
Aastha Khatiwada
}
