% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyGPATree.R
\name{ShinyGPATree}
\alias{ShinyGPATree}
\title{Run ShinyGPATree App}
\usage{
ShinyGPATree(object)
}
\arguments{
\item{object}{An object of class GPATree.}
}
\value{
Output of GPA-Tree model.
}
\description{
This function will initialize the ShinyGPATree App for dynamic and interactive visualization of GPA-Tree model results.
}
\examples{
\dontrun{
library(GPATree)

# load GPATree example data
data(GPATreeExampleData)

#fitting the GPATree model
fit <- GPATree(GPATreeExampleData$gwasPval, GPATreeExampleData$annMat)

# initialize the ShinyGPATree app
ShinyGPATree(fit)
}
}
\author{
Aastha Khatiwada
}
