% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{assoc}
\alias{assoc}
\alias{assoc,GPATree-method}
\title{Association mapping}
\usage{
\S4method{assoc}{GPATree}(object, FDR = 0.05, fdrControl = "global")
}
\arguments{
\item{object}{An object of class GPATree.}

\item{FDR}{FDR level. Value has to be between 0 and 1.}

\item{fdrControl}{Method to control FDR. Possible values are "global" (global FDR control) and "local" (local FDR control).}
}
\value{
Returns a MX2 matrix where the row represents SNPs, the fist column indicates the association between each SNP and phenotype, and the second column indicates the leaf in which the SNP falls
}
\description{
This function will implement association mapping for the GPA-Tree model.
}
\examples{
\dontrun{
library(GPATree)

# load GPATree example data
data(GPATreeExampleData)

#fitting the GPATree model
fit <- GPATree(GPATreeExampleData$gwasPval, GPATreeExampleData$annMat)

# pruning the GPATree model fit
assoc.fit <- assoc(fit, FDR = 0.05, fdrControl = "global")
}
}
\author{
Aastha Khatiwada
}
