% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyGPATree.R
\name{ShinyGPATree}
\alias{ShinyGPATree}
\title{Run ShinyGPATree App for dynamic and interactive}
\usage{
ShinyGPATree(object)
}
\arguments{
\item{object}{An object of class GPATree.}
}
\value{
Output of GPA-Tree model.
}
\description{
This function will run the shinyGPATree App.
}
\author{
Aastha Khatiwada
}
